Migrate integration tests to Test Kitchen
This commit eliminates the existing BATS integration tests and instead
introduces integration tests with Test Kitchen using the
kitchen-terraform and inspec-gcp plugins. The changes made to the
examples were to ensure that all scenarios could be run in parallel, to
correct syntax errors, and to use locals to make testing easier.