Updates for integration test CI pipeline
This commit includes changes necessary for enabling the integration
testing CI pipeline being defined in Concourse. The Makefile has been
updated to reference `test/ci_integration.sh` from the `make
test_integration` target which will export the correct Cloud SDK
environment variables based on environment variables coming out of the
CI pipeline (namely `$SERVICE_ACCOUNT_JSON` and `$PROJECT_ID`), handle
teardown of the environment correctly, and allow for the exit code of
the verify step to be communicated back to CI (instead of the exit code
from the destroy step). Also included is the
`test/fixtures/simulated_ci_environment` scenario which can be used to
setup a project and the same dependencies that are established by the CI
pipeline. This will allow a module consumer to be able to test locally
(or via the docker container) in the same way that CI testing is
happening remotely.