Update Makefile(s) for Developer Experience 2.0
The Developer Experience 2.0 changes have been focused on consolidating
as much code as possible into a new Docker image and reducing the code
duplication across the Terraform modules. Included in this task are
changes to `Makefile` designed to streamline the number of make targets
needed for local development and testing.