feat: removes the secondary_ranges map and adds it as a key to subnets
This commit removes the secondary_ranges map and makes it a required key in the subnets data structure. Although you now have to pass an empty list if your don't want to create a secondary range, this does allow all the configuration of a subnet to be in the same block.