fix: had to make subnet_flow_logs, subnet_private_access required inputs
I wasn't able to make these optional, the lookup function when used in the count.index loop, for whatever reason always returned the default value, even when the keys/value was passed as an input to the module. There may be another way to test for this, but I wasn't able to get any of them working. I tried differnt variations of "if has_key, then.." but wasn't able to get it working. Making these parameters required changes the API, making it more explicit (and longer) but has the advantage of actually working, and passing tests :smile: